+++ lib/src/fetch/FetchResponsePartition.h
@@ -49,7 +49,7 @@ class FetchResponsePartition : public WireFormatter, public PacketWriter, public
     MessageSet *messageSet;
 
     FetchResponsePartition(Packet *packet);
-    FetchResponsePartition(int partition, short int errorCode, long int highwaterMarkOffset, int messageSetSize, MessageSet *messageSet, bool releaseArrays = false);
+    FetchResponsePartition(int partition, short int errorCode, long long highwaterMarkOffset, int messageSetSize, MessageSet *messageSet, bool releaseArrays = false);
     ~FetchResponsePartition();
 
     unsigned char* toWireFormat(bool updatePacketSize = true);
